[REF-6108] open text files using utf-8 encoding
When opening file in text mode, default encoding is platform dependent
and defined by locale.getpreferredencoding(). If default is ascii,
decoding fails on unicode symbols.

It is recommended to explicitly use utf-8 when opening a file (or open
file in binary mode and then decode)

Python2 open() method doesn't support encoding argument, so it is
recommended to use io.open().