[PIPE-10] Decouple 'mlpiper' ('mlcomp') from 'mlops' module
By default a user will be able to run 'mlpiper' pipelines without
'mlops' module. The 'mlops' module was used for interaction
with ParallelM MCenter system, such as statistics reporting.

This work includes fixes at the framework execution layer
as well as in the unit-tests. In case the 'mlops' module
cannot be loaded, all of the code that uses the mlops library
is skipped.